Add gift message functionality and project structure
- Updated `.gitignore` to exclude `MigrationBackup/`.
- Introduced `Nop.Core.Domain.Customers` namespace with `NopCustomerDefaults` class for gift message attributes.
- Added `GiftMessage` property to `Order` class and related logic in `OrderProcessingService`.
- Created new project files: `Nop.Core.Bamboo.projitems` and `Nop.Services.Bamboo.projitems`.
- Implemented `GetAllProductAttributesExtendedAsync` in `IProductAttributeService`.
- Added `ApplyGiftMessageAsync` method in `CustomerService` and updated `ICustomerService` interface.
- Created API controllers for order management and token generation.
- Updated view files to display gift messages in order details and shopping cart.
- Created migration class to add `GiftMessage` column to the `Order` table.
- Updated project files and solution structure to include new Bamboo projects.

+++ Plugins/Nop.Plugin.DiscountRules.CustomDiscounts/Models/ConfigurationModel.cs
@@ -19,7 +19,7 @@ public record ConfigurationModel : BaseNopModel
     public bool Enabled_OverrideForStore { get; set; }
 
     [NopResourceDisplayName("Plugins.DiscountRules.CustomDiscounts.Fields.DiscountPercentage")]
-    public double DiscountPercentage { get; set; }
+    public decimal DiscountPercentage { get; set; }
     public bool DiscountPercentage_OverrideForStore { get; set; }
     #endregion
 }
